Refactor macOS-GodotAppView and macOS-GodotWindow to share core code
